requirements.txt
chromadb>=0.4.22
pysqlite3-binary>=0.5.0
llama-index-core>=0.10.0
# Pour chunking sémantique avec LlamaIndex :
# llama-index-embeddings-huggingface>=0.1.0